server/discovery-service/requirements.txt
fastapi==0.115.11
pydantic_core==2.27.2
uvicorn==0.34.0
python-consul2 ==0.1.5